Rename RootStackParamList.tsx to NavigationConfig.tsx
I will need to split the list in two for split view, and I'd like to put linking config in the same file since they are closely related
