+++ src/screens/LoginScreen.tsx
@@ -24,7 +24,7 @@ import { useCredentials } from "../credentials";
 import LinkButton from "../widgets/LinkButton";
 import { useNavigation } from "@react-navigation/native";
 import { StackNavigationProp } from "@react-navigation/stack";
-import { RootStackParamList } from "../RootStackParamList";
+import { RootStackParamList } from "../NavigationConfig";
 
 type NavigationProp = StackNavigationProp<RootStackParamList, "Login">;
